requirements.txt
colorama
requests
PyYAML
protobuf
clint
